ADDS testing fixture folder + sample staking pools fixture
